Split up platform tests
