+++ src/test/platform/initializeEffects_test.c
@@ -1,0 +1,7 @@
+#include "../test.h"
+
+void test_initializeEffects() {
+  if (verbose) {
+    safe_printf("TODO\n");
+  }
+}
