+++ src/test/platform/sendToApp_test.c
@@ -1,0 +1,7 @@
+#include "../test.h"
+
+void test_sendToApp() {
+  if (verbose) {
+    safe_printf("TODO\n");
+  }
+}
